# perf(crypto): bcrypto