--- a/packages/crypto/src/crypto/hash-algorithms.ts
+++ b/packages/crypto/src/crypto/hash-algorithms.ts
@@ -5,6 +5,7 @@ export class HashAlgorithms {
      * Create a "ripemd160" buffer.
      */
     public static ripemd160(buffer: Buffer | string): Buffer {
+        buffer = buffer instanceof Buffer ? buffer : Buffer.from(buffer);
         return RIPEMD160.digest(buffer);
     }
 
@@ -12,6 +13,7 @@ export class HashAlgorithms {
      * Create a "sha1" buffer.
      */
     public static sha1(buffer: Buffer | string): Buffer {
+        buffer = buffer instanceof Buffer ? buffer : Buffer.from(buffer);
         return SHA1.digest(buffer);
     }
 
@@ -20,7 +22,18 @@ export class HashAlgorithms {
      * @param  {Buffer} buffer
      * @return {Buffer}
      */
-    public static sha256(buffer: Buffer | string): Buffer {
+    public static sha256(buffer: Buffer | string | Buffer[]): Buffer {
+        if (Array.isArray(buffer)) {
+            let sha256 = SHA256.ctx;
+            sha256.init();
+            buffer.forEach(element => {
+                sha256 = sha256.update(element);
+            });
+
+            return sha256.final();
+        }
+
+        buffer = buffer instanceof Buffer ? buffer : Buffer.from(buffer);
         return SHA256.digest(buffer);
     }
 
@@ -28,6 +41,7 @@ export class HashAlgorithms {
      * Create a "hash160" buffer.
      */
     public static hash160(buffer: Buffer | string): Buffer {
+        buffer = buffer instanceof Buffer ? buffer : Buffer.from(buffer);
         return Hash160.digest(buffer);
     }
 
@@ -35,6 +49,7 @@ export class HashAlgorithms {
      * Create a "hash256" buffer.
      */
     public static hash256(buffer: Buffer | string): Buffer {
+        buffer = buffer instanceof Buffer ? buffer : Buffer.from(buffer);
         return Hash256.digest(buffer);
     }
 }