## Proposed changes
<!--
Describe the big picture of your changes here to communicate to the maintainers why we should accept this pull request. If it fixes a bug or resolves a feature request, be sure to link to that issue.
-->
Integrates [bcrypto](https://github.com/bcoin-org/bcrypto) which provides faster hashing functions than Node.
```
$ node benchmark/index.js
  HashAlgorithms
    ✓  bcrypto.sha256 x 1,124,900 ops/sec ±3.78% (80 runs sampled)
    ✓  node.sha256 x 454,155 ops/sec ±5.11% (72 runs sampled)
    ✓  bcrypto.sha1 x 1,527,517 ops/sec ±2.34% (84 runs sampled)
    ✓  node.sha1 x 482,688 ops/sec ±5.00% (70 runs sampled)
    ✓  bcrypto.ripemd160 x 955,411 ops/sec ±2.45% (84 runs sampled)
    ✓  node.ripemd160 x 399,268 ops/sec ±4.99% (71 runs sampled)
    ✓  bcrypto.hash160 x 879,670 ops/sec ±2.90% (79 runs sampled)
    ✓  node.hash160 x 200,743 ops/sec ±6.02% (64 runs sampled)
    ✓  bcrypto.hash256 x 993,016 ops/sec ±2.54% (84 runs sampled)
    ✓  node.hash256 x 268,063 ops/sec ±2.12% (76 runs sampled)
```

Resolves #1569

## Types of changes
<!--
What types of changes does your code introduce?
_Put an `x` in the boxes that apply and remove the rest of them: keep this PR as concise, but descriptive, as possible.._
-->

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [X] Refactoring (improve a current implementation without adding a new feature or fixing a bug)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Build (changes that affect the build system)
- [ ] Docs (documentation only changes)
- [ ] Test (adding missing tests or fixing existing tests)
- [ ] Other... Please describe:

## Checklist
<!--
_Put an `x` in the boxes that apply and remove this text and the rest of them. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your code._
-->

- [ ] I have read the [CONTRIBUTING](https://docs.ark.io/guidebook/contribution-guidelines/contributing.html) documentation
- [ ] Lint and unit tests pass locally with my changes
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)
<!--

## Further comments

If this is a relatively large or complex change, kick off the discussion by explaining why you chose the solution you did and what alternatives you considered, etc...
-->